Moving single classes one level up
