+++ stisim/__init__.py
@@ -1,14 +1,26 @@
-from .core.analyzers  import *
-from .core.connectors import *
-from .core.distributions import *
-from .core.interventions import *
-from .core.modules    import *
-from .core.parameters import *
-from .core.people     import *
-from .core.random     import *
-from .core.results    import *
-from .core.sim        import *
-from .core.demographics   import *
+# from .core.analyzers  import *
+# from .core.connectors import *
+# from .core.distributions import *
+# from .core.interventions import *
+# from .core.modules    import *
+# from .core.parameters import *
+# from .core.people     import *
+# from .core.random     import *
+# from .core.results    import *
+# from .core.sim        import *
+# from .core.demographics   import *
+
+from .analyzers  import *
+from .connectors import *
+from .distributions import *
+from .interventions import *
+from .modules    import *
+from .parameters import *
+from .people     import *
+from .random     import *
+from .results    import *
+from .sim        import *
+from .demographics   import *
 from .states.states   import *
 from .states.dinamicview import *
 from .states.fussedarray import *
