--- stisim/core/__init__.py
@@ -1,13 +1,0 @@
-from .modules import *
-# from .analyzers import *
-from .connectors import *
-from .distributions import *
-from .interventions import *
-from .modules import *
-from .parameters import *
-from .people import *
-from .random import *
-from .results import *
-from .sim import *
-from .demographics import *
-
