+++ stisim/parameters.py
