+++ stisim/random.py
