+++ stisim/results.py
