+++ stisim/sim.py
