+++ stisim/utils/logger.py
@@ -1,6 +1,3 @@
-"""
-Base classes for diseases
-"""
 
 import numpy as np
 import sciris as sc
